Generalize $package cache: remove VSP-specific naming from constants, helpers, and tests
