+++ vsm-app/worker/VSPPackageQueue.ts
@@ -1,183 +1,13 @@
 import { QUEUE_REDIS_URL } from '@/config'
 import Cache from '@/cache'
-import { JOB_STATUS, VSP_PACKAGE_CACHE_IDENTIFIER_SYSTEM, VSP_PACKAGE_CACHE_TAG } from '@/constants'
+import { JOB_STATUS } from '@/constants'
 import FhirClient from '@/backend/clients/FhirCdrClient'
 import Logger from '@/helpers/server/logger'
 import Queue from 'bull'
-import { createHash } from 'crypto'
 import { tsCredentialService } from '@/backend/services/TsCredentialService'
+import { hashPackageParams, searchCachedPackage, getCachedPackageContent, stripCacheMetadata, storeCachedPackage } from './packageCacheHelpers'
 
 const VSPPackageQueue = new Queue('vspPackage', QUEUE_REDIS_URL)
-
-/**
- * Produce a short hash of the $package parameters that affect output content.
- * If the terminology routes or endpoints change, the hash changes and the
- * previous cache entry becomes a miss.
- */
-function hashPackageParams(artifactRoute: string, artifactEndpointAddress: string, terminologyEndpointAddress: string): string {
-  const input = [artifactRoute, artifactEndpointAddress, terminologyEndpointAddress].join('|')
-  return createHash('sha256').update(input).digest('hex').slice(0, 12)
-}
-
-/**
- * Search for a cached $package Bundle by VSP identifier.
- * Returns the cached Bundle's ID or null if not found.
- */
-async function searchCachedPackage(vspId: string, vspVersion: string, paramsHash: string): Promise<string | null> {
-  try {
-    const { fetch: f } = require('undici')
-    const baseUrl = FhirClient.getInstance().baseUrl
-    const identifierParam = `${VSP_PACKAGE_CACHE_IDENTIFIER_SYSTEM}|${vspId}|${vspVersion}|${paramsHash}`
-    const url = `${baseUrl}/Bundle?identifier=${encodeURIComponent(identifierParam)}&_summary=true&_count=1`
-
-    const response = await f(url, {
-      method: 'GET',
-      headers: {
-        'Accept': 'application/fhir+json',
-        ...FhirClient.getInstance().customHeaders
-      }
-    })
-
-    if (!response.ok) {
-      Logger.getLogger().warn(`Cache search returned status ${response.status}`)
-      return null
-    }
-
-    const searchBundle = await response.json() as fhir4.Bundle
-    const entry = searchBundle.entry?.[0]
-    if (entry?.resource?.id) {
-      return entry.resource.id
-    }
-    return null
-  } catch (error: any) {
-    Logger.getLogger().warn(`Cache search failed: ${error.message || error}`)
-    return null
-  }
-}
-
-/**
- * Retrieve the full cached Bundle content by ID in the desired format.
- */
-async function getCachedPackageContent(bundleId: string, isJson: boolean): Promise<string | null> {
-  try {
-    const { fetch: f } = require('undici')
-    const baseUrl = FhirClient.getInstance().baseUrl
-    const url = `${baseUrl}/Bundle/${bundleId}`
-    const acceptHeader = isJson ? 'application/fhir+json' : 'application/fhir+xml'
-
-    const response = await f(url, {
-      method: 'GET',
-      headers: {
-        'Accept': acceptHeader,
-        ...FhirClient.getInstance().customHeaders
-      }
-    })
-
-    if (!response.ok) {
-      Logger.getLogger().warn(`Cache read returned status ${response.status}`)
-      return null
-    }
-
-    return await response.text()
-  } catch (error: any) {
-    Logger.getLogger().warn(`Cache read failed: ${error.message || error}`)
-    return null
-  }
-}
-
-/**
- * Strip cache metadata from a cached Bundle to reproduce original $package output.
- */
-function stripCacheMetadata(content: string, isJson: boolean): string {
-  if (isJson) {
-    const bundle = JSON.parse(content)
-    delete bundle.id
-    delete bundle.identifier
-    delete bundle.meta
-    return JSON.stringify(bundle)
-  } else {
-    // Remove <id .../> element
-    let result = content.replace(/<id[^>]*\/>\s*/m, '')
-    // Remove <identifier>...</identifier> block
-    result = result.replace(/<identifier>[\s\S]*?<\/identifier>\s*/m, '')
-    // Remove <meta>...</meta> block
-    result = result.replace(/<meta>[\s\S]*?<\/meta>\s*/m, '')
-    return result
-  }
-}
-
-/**
- * Store a $package result in the CDR as a cached Bundle.
- * Non-fatal — logs a warning if storage fails (e.g., Bundle too large).
- */
-async function storeCachedPackage(packageContent: string, vspId: string, vspVersion: string, paramsHash: string, isJson: boolean): Promise<void> {
-  try {
-    const { fetch: f } = require('undici')
-    const baseUrl = FhirClient.getInstance().baseUrl
-    const url = `${baseUrl}/Bundle`
-
-    const cacheIdentifier = {
-      system: VSP_PACKAGE_CACHE_IDENTIFIER_SYSTEM,
-      value: `${vspId}|${vspVersion}|${paramsHash}`
-    }
-
-    let body: string
-    let contentType: string
-
-    if (isJson) {
-      const bundle = JSON.parse(packageContent)
-      bundle.identifier = cacheIdentifier
-      bundle.meta = {
-        ...bundle.meta,
-        tag: [...(bundle.meta?.tag || []), VSP_PACKAGE_CACHE_TAG]
-      }
-      body = JSON.stringify(bundle)
-      contentType = 'application/fhir+json'
-    } else {
-      // Inject <identifier> and <meta> XML elements after the <Bundle ...> opening tag
-      const identifierXml = `<identifier><system value="${cacheIdentifier.system}"/><value value="${cacheIdentifier.value}"/></identifier>`
-      const metaXml = `<meta><tag><system value="${VSP_PACKAGE_CACHE_TAG.system}"/><code value="${VSP_PACKAGE_CACHE_TAG.code}"/></tag></meta>`
-
-      // Insert after the <Bundle ...> opening tag, before the first child element
-      body = packageContent.replace(
-        /(<Bundle[^>]*>)/,
-        `$1${identifierXml}${metaXml}`
-      )
-      contentType = 'application/fhir+xml'
-    }
-
-    const response = await f(url, {
-      method: 'POST',
-      body,
-      headers: {
-        'Content-Type': contentType,
-        'Accept': 'application/fhir+json',
-        ...FhirClient.getInstance().customHeaders
-      }
-    })
-
-    if (!response.ok) {
-      const errorText = await response.text()
-      Logger.getLogger().warn(
-        `Failed to cache $package result for VSP ${vspId} (${packageContent.length} bytes).\n` +
-        `This is non-fatal — subsequent exports will re-run $package.\n` +
-        `To enable caching for large packages, increase the FHIR server's\n` +
-        `maximum resource size limit.\n` +
-        `Error: ${errorText}`
-      )
-    } else {
-      Logger.getLogger().info(`Cached $package result for VSP ${vspId} (${packageContent.length} bytes)`)
-    }
-  } catch (error: any) {
-    Logger.getLogger().warn(
-      `Failed to cache $package result for VSP ${vspId} (${packageContent.length} bytes).\n` +
-      `This is non-fatal — subsequent exports will re-run $package.\n` +
-      `To enable caching for large packages, increase the FHIR server's\n` +
-      `maximum resource size limit.\n` +
-      `Error: ${error.message || error}`
-    )
-  }
-}
 
 interface VSPPackageJobData {
   vspId: string
@@ -427,5 +257,4 @@ VSPPackageQueue.process(async function (job: any, done) {
   }
 })
 
-export { hashPackageParams, stripCacheMetadata }
 export default VSPPackageQueue
